refactor(bridge): colocate LayerSwap fee types with chain types
Move LayerSwap-specific fee estimation types into per-chain types.ts
files (matching the CCTP/OFT pattern) and drop the concrete-class
exports from the bridge barrel. The Ethereum variant intersects the
shared Ethereum fee base; Solana variants are listed directly in the
SDK-level BridgeDepositFeeEstimation union since they share no base.
