--- src/bridge/ethereum/layerswap/index.ts
@@ -1,24 +1,0 @@
-export {
-  LayerSwapBridge,
-  type LayerSwapDepositFeeEstimation,
-} from "@/bridge/ethereum/layerswap/LayerSwapBridge";
-export { LayerSwapApi } from "@/bridge/ethereum/layerswap/LayerSwapApi";
-export {
-  LayerSwapApiError,
-  type LayerSwapApiConfig,
-  type LayerSwapCreateRequest,
-  type LayerSwapQuoteRequest,
-  type LsApiResponse,
-  type LsDepositAction,
-  type LsLimits,
-  type LsNetwork,
-  type LsNetworkType,
-  type LsQuote,
-  type LsQuoteResponse,
-  type LsRoute,
-  type LsSwap,
-  type LsSwapResponse,
-  type LsSwapStatus,
-  type LsToken,
-  type LsTransaction,
-} from "@/bridge/ethereum/layerswap/types";
