+++ src/bridge/ethereum/types.ts
@@ -41,3 +41,13 @@ export type OftDepositFeeEstimation = EthereumDepositFeeEstimation & {
   /** LayerZero interchain fee (in ETH, included in msg.value of the deposit tx). */
   interchainFee: Amount;
 };
+
+/** Fee estimation returned by LayerSwap, extending the base Ethereum fee shape. */
+export type LayerSwapDepositFeeEstimation = EthereumDepositFeeEstimation & {
+  /** LayerSwap service fee portion. */
+  serviceFee: Amount;
+  /** Amount the recipient will receive after fees. */
+  receiveAmount: Amount;
+  /** Estimated completion time (e.g. "00:02:00"). */
+  avgCompletionTime: string;
+};
