+++ src/bridge/index.ts
@@ -2,9 +2,4 @@ export * from "@/bridge/tokens/repository";
 export * from "@/bridge/operator";
 export * from "@/bridge/ethereum";
 export * from "@/bridge/solana/types";
-export * from "@/bridge/ethereum/layerswap";
-export {
-  SolanaLayerSwapBridge,
-  type SolanaLayerSwapDepositFeeEstimation,
-} from "@/bridge/solana/SolanaLayerSwapBridge";
 export type { BridgeDepositOptions } from "@/bridge/types/BridgeInterface";
